# Optimize a slew of Cargo internals