… Cargo.lock
… src/cargo/core/source.rs
… src/cargo/core/summary.rs
… src/cargo/ops/cargo_clean.rs
… src/cargo/ops/cargo_install.rs
… src/cargo/ops/cargo_rustc/context.rs
… src/cargo/ops/cargo_rustc/fingerprint.rs
… src/cargo/ops/cargo_rustc/job_queue.rs
… src/cargo/ops/cargo_rustc/mod.rs
… src/cargo/ops/cargo_rustc/output_depinfo.rs
… src/cargo/ops/resolve.rs
… src/cargo/sources/directory.rs
… src/cargo/sources/git/source.rs
… src/cargo/sources/path.rs
… src/cargo/sources/registry/index.rs
… src/cargo/sources/registry/local.rs
… src/cargo/sources/registry/mod.rs
… src/cargo/sources/registry/remote.rs
… src/cargo/sources/replaced.rs
… src/cargo/util/graph.rs
… src/cargo/util/paths.rs
… src/cargo/util/toml.rs
… tests/resolve.rs