@alexcrichton 2017-06-03
> 👍 (at least, the branch is named hamt :) )

You can see how I started down this rabbit hole!