@Mark-Simulacrum 2017-06-05
Failure looks legitimate.

```
error: no method named `query` found for type `std::vec::Vec<core::summary::Summary>` in the current scope
   --> src/cargo/core/registry.rs:444:32
    |
444 |                 self.summaries.query(dep, f)
    |                                ^^^^^
    |
    = help: items from traits can only be used if the trait is im
```